+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.7.8 (WIP)
+## 0.7.8 (2021-04-29)
 
 * Add some Terser workarounds
 * Make `String#replaceAll()` a polyfill, as it is now part of ECMAScript-2021